requirements.txt
Django==2.0.2
django-mssql==1.8
djangorestframework==3.7.7
pytz==2018.3
